Update back buttons UI placement
