+++ client/src/routes/settings.tsx
@@ -1,9 +1,11 @@
+import { SidebarBack } from "@/components/settings/BackButtons";
 import { Button } from "@/components/ui/button";
 import {
   Sidebar,
   SidebarGroup,
   SidebarGroupContent,
   SidebarGroupLabel,
+  SidebarHeader,
   SidebarMenu,
   SidebarMenuButton,
   SidebarMenuItem,
@@ -12,7 +14,7 @@ import {
 } from "@/components/ui/sidebar";
 import { authClient } from "@/lib/auth-client";
 import { createFileRoute, Link, Outlet, useRouter } from "@tanstack/react-router";
-import { ChevronLeft, ChevronRight, Key, User, UserIcon } from "lucide-react";
+import { ArrowLeft, ChevronLeft, ChevronRight, Key, User, UserIcon } from "lucide-react";
 import React from "react";
 
 export const Route = createFileRoute("/settings")({
@@ -40,35 +42,27 @@ function SidebarComponent() {
     }
   }, [sidebar.isMobile]);
 
-  function exitSidebar() {
-    if (router.history.canGoBack()) {
-      router.history.back();
-    } else {
-      router.navigate({ to: "/" });
-    }
-  }
-
   return (
     <Sidebar>
-      {sidebar.isMobile ? (
-        <div className="flex items-center justify-between w-full mt-3">
-          <Button variant="ghost" onClick={exitSidebar}>
-            <ChevronLeft className="size-6" />
-          </Button>
-
-          <h1 className="absolute left-1/2 transform -translate-x-1/2 font-bold text-lg">Settings</h1>
-        </div>
-      ) : null}
+      <SidebarHeader>
+        <SidebarBack />
+      </SidebarHeader>
 
       <SidebarGroup>
         <SidebarGroupLabel>User Settings</SidebarGroupLabel>
         <SidebarGroupContent>
           <SidebarMenu>
             <SettingsMenuButton>
-              <Link to="/settings" replace><User />General</Link>
+              <Link to="/settings" replace>
+                <User />
+                General
+              </Link>
             </SettingsMenuButton>
             <SettingsMenuButton>
-              <Link to="/settings/keys" replace><Key />Key Management</Link>
+              <Link to="/settings/keys" replace>
+                <Key />
+                Key Management
+              </Link>
             </SettingsMenuButton>
           </SidebarMenu>
         </SidebarGroupContent>
@@ -79,10 +73,16 @@ function SidebarComponent() {
         <SidebarGroupContent>
           <SidebarMenu>
             <SettingsMenuButton>
-              <Link to="/settings" replace><User />General</Link>
+              <Link to="/settings" replace>
+                <User />
+                General
+              </Link>
             </SettingsMenuButton>
             <SettingsMenuButton>
-              <Link to="/settings/keys" replace><Key />Key Management</Link>
+              <Link to="/settings/keys" replace>
+                <Key />
+                Key Management
+              </Link>
             </SettingsMenuButton>
           </SidebarMenu>
         </SidebarGroupContent>
@@ -96,7 +96,12 @@ function SettingsMenuButton({ children }: { children: React.ReactNode }) {
 
   return (
     <SidebarMenuItem className="relative flex items-center">
-      <SidebarMenuButton asChild children={children} className="text-lg font-semibold z-10" onClick={() => sidebar.setOpenMobile(false)} />
+      <SidebarMenuButton
+        asChild
+        children={children}
+        className="text-lg font-semibold z-10"
+        onClick={() => sidebar.setOpenMobile(false)}
+      />
       {sidebar.isMobile ? <ChevronRight className="absolute right-0 h-full" /> : null}
     </SidebarMenuItem>
   );
